+++ notebooks/dum/dummy.clj
@@ -3,17 +3,26 @@
             [tablecloth.api :as tc]))
 
 (note-to-test/define-value-representations!
-  [{:predicate var?
-    :representation (constantly :var)}
+  [{:predicate symbol?
+    :representation (partial str "symbol ")}
    {:predicate tc/dataset?
     :representation (fn [ds]
                       (-> ds
                           (update-vals vec)
                           (->> (into {}))))}
    {:predicate (partial = 5)
-    :representation (constantly :five)}])
+    :representation (constantly :five)}
+   {:predicate map?
+    :representation (fn [m]
+                      (-> m
+                          (update-keys note-to-test/represent-value)
+                          (update-vals note-to-test/represent-value)))}
+   {:predicate sequential?
+    :representation (partial mapv note-to-test/represent-value)}])
 
 (+ 2 3)
+
+{:x (+ 2 3)}
 
 (+ 4
    5
@@ -27,14 +36,11 @@
 
 (f 12)
 
-(require 'clojure.java.io)
-
 (-> {:x [1 2 3]}
     tc/dataset
     (tc/map-columns :y [:x] (partial * 10)))
 
 (f 13)
-
 
 {:x 9}