+++ test/dum/dummy_generated_test.clj
@@ -2,58 +2,64 @@
   (:require
    [clojure.test :refer [deftest is]]
    [scicloj.note-to-test.v1.api :as note-to-test]
-   [tablecloth.api :as tc]
-   clojure.java.io))
+   [tablecloth.api :as tc]))
 
 (deftest test-everything
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           (note-to-test/define-value-representations!
-            [{:predicate var?
-              :representation (constantly :var)}
+            [{:predicate symbol?
+              :representation (partial str "symbol ")}
              {:predicate tc/dataset?
               :representation (fn [ds]
                                 (-> ds
                                     (update-vals vec)
                                     (->> (into {}))))}
              {:predicate (partial = 5)
-              :representation (constantly :five)}]))
-       [:ok]))
+              :representation (constantly :five)}
+             {:predicate map?
+              :representation (fn [m]
+                                (-> m
+                                    (update-keys note-to-test/represent-value)
+                                    (update-vals note-to-test/represent-value)))}
+             {:predicate sequential?
+              :representation (partial mapv note-to-test/represent-value)}]))
+       {:value [:ok], :meta nil}))
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           (+ 2 3))
-       :five))
+       {:value :five, :meta nil}))
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
+          {:x (+ 2 3)})
+       {:value {:x :five}, :meta nil}))
+
+  (is (= (note-to-test/represent-value-with-meta
           (+ 4
              5
              6))
-       15))
+       {:value 15, :meta nil}))
 
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           (defn f [x]
             (+ x 19)))
-       :var))
+       [:var]))
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           (f 12))
-       31))
+       {:value 31, :meta nil}))
 
-  (is (= (note-to-test/represent-value
-          (require 'clojure.java.io))
-       nil))
-
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           (-> {:x [1 2 3]}
               tc/dataset
               (tc/map-columns :y [:x] (partial * 10))))
-       {:x [1 2 3], :y [10 20 30]}))
+       {:value {:x [1 2 3], :y [10 20 30]}, :meta {:name "_unnamed"}}))
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           (f 13))
-       32))
+       {:value 32, :meta nil}))
 
-  (is (= (note-to-test/represent-value
+  (is (= (note-to-test/represent-value-with-meta
           {:x 9})
-       {:x 9})))
+       {:value {:x 9}, :meta nil})))